refactor: dedupe code
