+++ packages/cli/src/react/jsx/utils/stringParsing/processTranslationCall/handleDeriveTranslationCall.ts
@@ -2,20 +2,8 @@ import { NodePath } from '@babel/traverse';
 import * as t from '@babel/types';
 import { ParsingConfig } from '../types.js';
 import { ParsingOutput } from '../types.js';
-import { handleDeriveExpression } from '../derivation/parseDerive.js';
-import { nodeToStrings } from '../../parseString.js';
-import { indexVars } from 'generaltranslation/internal';
-import { isValidIcu } from '../../../evaluateJsx.js';
-import {
-  warnInvalidIcuSync,
-  warnNonStringSync,
-} from '../../../../../console/index.js';
-import generateModule from '@babel/generator';
-import { randomUUID } from 'node:crypto';
+import { deriveExpression } from '../derivation/index.js';
 import { InlineMetadata } from './extractStringEntryMetadata.js';
-
-// Handle CommonJS/ESM interop
-const generate = generateModule.default || generateModule;
 
 /**
  * For the processTranslationCall function, this function handles the case where a string with derive is used.
@@ -41,57 +29,12 @@ export function handleDeriveTranslationCall({
   output: ParsingOutput;
   index?: number;
 }): void {
-  // parse derivable expression
-  const stringNode = handleDeriveExpression({
-    expr: arg,
+  deriveExpression({
     tPath,
-    file: config.file,
-    parsingOptions: config.parsingOptions,
-    errors: output.errors,
+    expr: arg,
+    metadata,
+    config,
+    output,
+    index,
   });
-
-  // Nothing returned, push error
-  if (!stringNode) {
-    output.errors.push(
-      warnNonStringSync(
-        config.file,
-        generate(arg).code,
-        `${arg.loc?.start?.line}:${arg.loc?.start?.column}`
-      )
-    );
-    return;
-  }
-
-  // validate ICU
-  const strings = nodeToStrings(stringNode).map(indexVars);
-  if (!config.ignoreInvalidIcu) {
-    for (const string of strings) {
-      const { isValid, error } = isValidIcu(string);
-      if (!isValid) {
-        output.warnings.add(
-          warnInvalidIcuSync(
-            config.file,
-            string,
-            error ?? 'Unknown error',
-            `${arg.loc?.start?.line}:${arg.loc?.start?.column}`
-          )
-        );
-        return;
-      }
-    }
-  }
-
-  const temporaryDeriveId = `derive-temp-id-${randomUUID()}`;
-  for (const string of strings) {
-    output.updates.push({
-      dataFormat: 'ICU',
-      source: string,
-      metadata: {
-        ...metadata,
-        // Add the index if an id and index is provided (for handling when registering an array of strings)
-        ...(metadata.id && index != null && { id: `${metadata.id}.${index}` }),
-        staticId: temporaryDeriveId,
-      },
-    });
-  }
 }
